requirement.txt
langchain-aws==0.1.2
qdrant-client==1.9.0
boto3==1.34.112
python-dotenv==1.0.1
markdown-it-py==3.0.0
langchain-community
langchain-qdrant